requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
firebase-admin>=6.2
python-dotenv>=1.0
django-cors-headers>=4.3
razorpay>=1.4
Pillow>=10.0
